requirements.txt
torch
torchvision
tqdm
Pillow
numpy
matplotlib
